requirements.txt
openai
pandas
pyyaml
click
rich
python-dotenv
requests